move example with annotations and test project to test fixtures
