+++ src/reducers/projects.js
@@ -19,7 +19,7 @@ import * as instanceMap from '../store/instanceMap';
 const initialState = {};
 
 if (process.env.NODE_ENV === 'test') {
-  const testProject = require('./testProject').project;
+  const testProject = require('../../test/res/testProject').project;
   Object.assign(initialState, {
     test: testProject,
   });
