+++ test/res/exampleWithAnnotations.js
